fixed packaging issue, refactored a bit of code; this module could really use some TLC...  had to restructure a bit too...
