+++ extensions/ezmsg-websocket/ezmsg/websocket/__version__.py
@@ -1,1 +1,1 @@
-__version__ = '1.0.1'
+__version__ = '1.1.0'
